fix: norm #70
