+++ srcs/fixtures.c
@@ -6,61 +6,11 @@
 /*   By: stan <stan@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/29 20:51:48 by stan              #+#    #+#             */
-/*   Updated: 2022/04/29 22:21:46 by stan             ###   ########.fr       */
+/*   Updated: 2022/04/29 22:25:37 by stan             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "main.h"
-
-void	verif_player_pos(t_game *game)
-{
-	int		i;
-	int		x;
-
-	i = 0;
-	while (game->map[i])
-	{
-		x = 0;
-		while (game->map[i][x])
-		{
-			if (game->map[i][x] == 'N'
-				|| game->map[i][x] == 'S'
-				|| game->map[i][x] == 'E'
-				|| game->map[i][x] == 'W')
-			{
-				game->player_y = i;
-				game->player_x = x;
-			}
-			x++;
-		}
-		i++;
-	}
-}
-
-char	verif_player_exist(t_game game)
-{
-	int		i;
-	int		x;
-
-	i = 0;
-	while (game.map[i])
-	{
-		x = 0;
-		while (game.map[i][x])
-		{
-			if (game.map[i][x] == 'N'
-				|| game.map[i][x] == 'S'
-				|| game.map[i][x] == 'E'
-				|| game.map[i][x] == 'W')
-				return (game.map[i][x]);
-			x++;
-		}
-		i++;
-	}
-	free_game(game);
-	printf("Error\nDid not found player on map\n");
-	exit(1);
-}
 
 void	verif_map_config(t_game game)
 {
@@ -91,31 +41,6 @@ void	verif_map_config(t_game game)
 	}
 }
 
-void	check_rgb(t_game game)
-{
-	int		err;
-
-	err = 0;
-	if (game.ceil_color.color1 > 255 || game.ceil_color.color1 < 0)
-		err = 1;
-	if (game.ceil_color.color2 > 255 || game.ceil_color.color2 < 0)
-		err = 1;
-	if (game.ceil_color.color3 > 255 || game.ceil_color.color3 < 0)
-		err = 1;
-	if (game.floor_color.color1 > 255 || game.floor_color.color1 < 0)
-		err = 1;
-	if (game.floor_color.color2 > 255 || game.floor_color.color2 < 0)
-		err = 1;
-	if (game.floor_color.color3 > 255 || game.floor_color.color3 < 0)
-		err = 1;
-	if (err)
-	{
-		free_game(game);
-		printf("Error\nRGB color should be between 0 and 255\n");
-		exit(1);
-	}
-}
-
 void	check_multiple_player(t_game game)
 {
 	int		i;
@@ -142,80 +67,6 @@ void	check_multiple_player(t_game game)
 		free_game(game);
 		printf("Error\nCannot have no more or less than one player in the map\n");
 		exit(1);
-	}
-}
-
-void	check_line_closed_top(t_game game, int i, int x)
-{
-	if (game.map[i][x] == ' ')
-	{
-		if (game.map[i + 1][x] == '0' || game.map[i + 1][x - 1] != '1' || game.map[i + 1][x + 1] != '1')
-		{
-			printf("Error\nThe map is not closed at position %d %d\n", i, x);
-			free_game(game);
-			exit(1);
-		}
-		if (game.map[i + 1][x] == ' ')
-		{
-			check_line_closed_top(game, i + 1, x);
-		}
-		if (game.map[i + 1][x] == '1')
-			return ;
-	}
-}
-
-void	check_line_closed_bottom(t_game game, int i, int x)
-{
-	if (game.map[i][x] == ' ')
-	{
-		if (game.map[i - 1][x] == '0' || game.map[i - 1][x - 1] != '1' || game.map[i - 1][x + 1] != '1')
-		{
-			printf("Error\nThe map is not closed at position %d %d\n", i, x);
-			free_game(game);
-			exit(1);
-		}
-		if (game.map[i - 1][x] == ' ')
-		{
-			check_line_closed_bottom(game, i - 1, x);
-		}
-		if (game.map[i - 1][x] == '1')
-			return ;
-	}
-}
-
-void	check_line_closed_right(t_game game, int i, int x)
-{
-	if (game.map[i][x + 1] != 0)
-		return ;
-	if (game.map[i][x] == ' ')
-	{
-		if (game.map[i + 1][x] == '0' || game.map[i - 1][x] == '0')
-		{
-			printf("Error\nThe map is not closed at position %d %d\n", i, x);
-			free_game(game);
-			exit(1);
-		}
-		if (game.map[i][x - 1] == ' ')
-			check_line_closed_right(game, i, x - 1);
-		if (game.map[i][x - 1] == '1')
-			return ;
-	}
-}
-
-void	check_line_closed_left(t_game game, int i, int x)
-{
-	if (game.map[i][x] == ' ')
-	{
-		if (game.map[i + 1][x] == '0' || game.map[i - 1][x] == '0')
-		{
-			printf("Error\nThe map is not closed at position %d %d\n", i, x);
-			free_game(game);
-			exit(1);
-		}
-		if (game.map[i][x + 1] == ' ')
-			check_line_closed_left(game, i, x + 1);
-		if (game.map[i][x + 1] == '1')
-			return ;
 	}
 }
