+++ srcs/main.c
@@ -6,7 +6,7 @@
 /*   By: stan <stan@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/29 20:21:03 by stan              #+#    #+#             */
-/*   Updated: 2022/04/29 19:14:18 by stan             ###   ########.fr       */
+/*   Updated: 2022/04/29 22:23:42 by stan             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@ int	main(int ac, char **av)
 	check_config_file(av[1]);
 	game = init_game(av[1]);
 	init_info(&game);
-	// init_mlx(game);
+	init_mlx(game);
 	free_game(game);
 	return (0);
 }
